requirements.txt
aniso8601==9.0.1
astroid==2.8.5
attrs==21.2.0
beautifulsoup4==4.10.0
blis==0.7.4
Bottleneck==1.3.2
brotlipy==0.7.0
catalogue==2.0.6
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
coverage==5.5
cryptography==35.0.0
cymem==2.0.5
en-core-web-sm-abd==0.1
flake8==3.9.2
Flask==2.0.2
Flask-RESTful==0.3.9
idna==3.2
importlib-metadata==4.8.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
jsonschema==3.0.2
langcodes==3.2.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
murmurhash==1.0.5
numexpr==2.7.3
numpy==1.21.2
packaging==21.0
pandas==1.3.4
pathy==0.6.1
pip==21.2.4
plac==1.1.0
platformdirs==2.4.0
pprintpp==0.4.0
preshed==3.0.5
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
pylint==2.11.1
pyOpenSSL==21.0.0
pyparsing==3.0.5
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.1
setuptools==58.0.4
shellingham==1.4.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.12
threadpoolctl==2.2.0
toml==0.10.2
tqdm==4.62.3
typer==0.4.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.2
wheel==0.37.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.13.3
zipp==3.6.0
python-dotenv
python-jose
flask-cors
six
